rename HIRT_Multichannel_Convolution/{ConvolveUtilities => utilities}.hpp
