+++ include/HIRT_Multichannel_Convolution/n_to_mono.hpp
@@ -2,7 +2,7 @@
 #ifndef CONVOLVE_N_TO_MONO_HPP
 #define CONVOLVE_N_TO_MONO_HPP
 
-#include "ConvolveUtilities.hpp"
+#include "utilities.hpp"
 #include "mono.hpp"
 
 #include <algorithm>
