+++ include/HIRT_Multichannel_Convolution/utilities.hpp
